seminars/08-balancing-sharding/kv/proxy/requirements.txt
Flask==2.3.3
requests==2.31.0
pyyaml==6.0.1
